+++ src/modules/search/services/index.ts
@@ -1,2 +1,3 @@
 export { KeywordSearchService } from './KeywordSearchService';
 export { SemanticSearchService } from './SemanticSearchService';
+export { HybridSearchService, applyRRF } from './HybridSearchService';